Add files via upload
Want to compare sales and orders using a report?
Which month has the most sales and orders?
Who will buy more in 2022; men or women?
What are the different orders for 2022?
The first 10 states on the list contributed to sales.
The relationship between age and gender as a value judgment.
Which tube helps generate the most sales? Best selling category? etc.